+++ apps/explorer/src/components/homepage/StatsCard.tsx
@@ -1,35 +1,9 @@
 "use client";
 
-import { useEffect, useState } from "react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { AnimatedNumber } from "../ui/AnimatedNumber";
 
-const AnimatedNumber = ({
-  end,
-  duration = 2000,
-}: {
-  end: number;
-  duration?: number;
-}) => {
-  const [count, setCount] = useState(0);
-
-  useEffect(() => {
-    let startTime: number | null = null;
-    const animate = (timestamp: number) => {
-      if (!startTime) startTime = timestamp;
-      const progress = timestamp - startTime;
-      const percentage = Math.min(progress / duration, 1);
-      setCount(Math.floor(end * percentage));
-      if (percentage < 1) {
-        requestAnimationFrame(animate);
-      }
-    };
-    requestAnimationFrame(animate);
-  }, [end, duration]);
-
-  return <span className="tabular-nums">{count.toLocaleString()}</span>;
-};
-
-export default function InfoCard() {
+export default function StatsCard() {
   return (
     <Card className="px-12 py-12">
       {/* <CardHeader>
