+++ apps/explorer/src/components/ui/AnimatedNumber.tsx
@@ -1,0 +1,27 @@
+import { useState, useEffect } from "react";
+
+export const AnimatedNumber = ({
+  end,
+  duration = 2000,
+}: {
+  end: number;
+  duration?: number;
+}) => {
+  const [count, setCount] = useState(0);
+
+  useEffect(() => {
+    let startTime: number | null = null;
+    const animate = (timestamp: number) => {
+      if (!startTime) startTime = timestamp;
+      const progress = timestamp - startTime;
+      const percentage = Math.min(progress / duration, 1);
+      setCount(Math.floor(end * percentage));
+      if (percentage < 1) {
+        requestAnimationFrame(animate);
+      }
+    };
+    requestAnimationFrame(animate);
+  }, [end, duration]);
+
+  return <span className="tabular-nums">{count.toLocaleString()}</span>;
+};
